video_copyright_checker/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
minio==7.2.7
moviepy==1.0.3
numpy==1.26.4
pika==1.3.2
pillow==10.3.0
proglog==0.1.10
pycparser==2.22
pycryptodome==3.20.0
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
datasets==2.19.2
PyAudio==0.2.14
pydub==0.25.1
scipy==1.13.1
pandas==2.2.2
torch --index-url https://download.pytorch.org/whl/cu121
accelerate==0.31.0
transformers==4.41.2
tokenizers==0.19.1
ffmpeg==1.4
